pyproject.toml
[tool.poetry]
name = "jpeg-qtable-simulator"
version = "0.1.0"
description = ""
authors = ["KOUNOIKE Yuusuke <kounoike.yuusuke@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyvirtualcam = "^0.8.0"
Pillow = "^8.3.2"
click = "^8.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
